+++ src/main/kotlin/graphMateKT/solutions/Day7Laboratories2025.kt
@@ -12,7 +12,7 @@ internal fun main() {
 }
 
 /** Solves https://open.kattis.com/problems/Day7Laboratories2025 */
-internal fun Day7Laboratories2025(input: List<String>): BigInteger {
+internal fun Day7Laboratories2025(input: List<String>): Long {
     val g = Grid(input)
     val sink = Tile(0,0, '^')
     g.addNode(sink)
